requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
anthropic>=0.3.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
httpx~=0.27.2
h11~=0.14.0
PySide6~=6.8.0.1
pillow~=11.0.0
pip~=24.0
distro~=1.9.0
sniffio~=1.3.1
typing_extensions~=4.12.2
click~=8.1.7
httpcore~=1.0.7
idna~=3.10
certifi~=2024.8.30
jiter~=0.7.1
pluggy~=1.5.0
iniconfig~=2.0.0
starlette~=0.41.2
pydantic_core~=2.23.4
anyio~=4.6.2.post1
annotated-types~=0.7.0
packaging~=24.2
setuptools~=65.5.0
async-timeout~=4.0.3